Pointset now uses ZincObject prototype. Move more functions to the ZincObject prototype.
